requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.100.0
python-dateutil==2.8.2
